requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
fastapi==0.115.6
uvicorn==0.32.1
websockets==13.1
telethon==1.37.0
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==11.0.0
celery==5.4.0
redis==5.2.1
channels==4.2.0
channels-redis==4.2.0
gunicorn==23.0.0
openai==1.95.1
requests==2.32.4
qrcode[pil]==8.2

